Fix bug: undefined method `select_all_column?' for []:ActiveRecord::Relation
If ActiveRecord::Relation is load prior to SecondLevelCache, the
SecondLevelCache::ActiveRecord::FinderMethods will only apply the methods
in included block, not all the methods including in the module.

The easiest solution is just to include the module to ActiveRecord::Relation
directly.

The way to run the test is

```ruby
rake test TEST=test/require_test.rb
```